requirements.txt
beautifulsoup4==4.12.3
pdfminer.six==20231228
requests==2.32.3
